+++ lib/domain/initial_setting/initial_setting_state_notifier.dart
@@ -12,7 +12,7 @@ import 'package:pilll/provider/pill_sheet_modified_history.dart';
 import 'package:pilll/provider/setting.dart';
 import 'package:pilll/service/auth.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
-import 'package:pilll/database/user.dart';
+import 'package:pilll/provider/user.dart';
 import 'package:pilll/util/datetime/day.dart';
 import 'package:riverpod/riverpod.dart';
